requirements.txt
blinker==1.4
click==7.1.2
Flask==1.1.2
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
SQLAlchemy==1.3.19
Werkzeug==1.0.1
WTForms==2.3.3
python-dotenv==0.14.0